Implemented the "reflow" command.
For example, if you have the following table in your source file:

Type  Message
Foo  I like bananas very much. Even so much that I adore them.

Position your cursor in it and call ReformatTable().  This makes the
following:

+======+===========================================================+
| Type | Message                                                   |
+======+===========================================================+
| Foo  | I like bananas very much. Even so much that I adore them. |
+------+-----------------------------------------------------------+

But if you don't like these kinds of long wrappings, you can now simply
remove (or add, for that matter) some of the '='-signs from the top row,
like this:

+======+================================+
| Type | Message                                                   |
+======+===========================================================+
| Foo  | I like bananas very much. Even so much that I adore them. |
+------+-----------------------------------------------------------+

And then, call ReflowTable():

+======+================================+
| Type | Message                        |
+======+================================+
| Foo  | I like bananas very much. Even |
|      | so much that I adore them.     |
+------+--------------------------------+

This fixed issue #1.